rustigram_types

Module file

File and media types.

Structs

Animation

Animation file (GIF or H.264/MPEG-4 AVC, no sound).

Audio

Audio file to be treated as music.

Document

General file (not photo, voice, audio, or video).

File

A file ready to be downloaded.

InputMediaAnimation

An animation (GIF or silent H.264) to include in a media group.

InputMediaAudio

An audio file to include in a media group.

InputMediaDocument

A document (general file) to include in a media group.

InputMediaLivePhoto

A live photo to be sent as part of a media group or to replace an existing media message.

InputMediaLocation

A location to be sent as poll media.

InputMediaPhoto

A photo to include in a media group or replace an existing media message.

InputMediaSticker

A sticker file to be sent as poll option media.

InputMediaVenue

A venue to be sent as poll media.

InputMediaVideo

A video to include in a media group or replace an existing media message.

InputMediaVoiceNote

A voice message file to be sent.

InputPaidMediaLivePhoto

A live photo to send as paid media.

InputPaidMediaPhoto

A photo to send as paid media.

InputPaidMediaVideo

A video to send as paid media.

InputProfilePhotoAnimated

An animated profile photo in MPEG4 format.

InputProfilePhotoStatic

A static profile photo in JPEG format.

LivePhoto

A photo with a short video (live photo).

PassportFile

Encrypted passport file.

PhotoSize

One size of a photo or file thumbnail.

Video

Video file.

VideoNote

Rounded-square MPEG4 video note.

VideoQuality

A video file of a specific quality.

Voice

Voice note (OGG/OPUS audio).

Enums

InputFile

Represents a file to be sent.

InputMedia

The media content of a message to be sent or edited.

InputPaidMedia

Paid media to send via sendPaidMedia.

InputProfilePhoto

A profile photo to set via setMyProfilePhoto or setBusinessAccountProfilePhoto.