rustigram_types

Module gifts

Gift types.

Structs

Gift

A Telegram gift that can be sent to users.

GiftBackground

Background of a regular gift.

GiftInfo

Service message: a regular gift was sent or received.

Gifts

Collection of available gifts returned by getAvailableGifts.

UniqueGift

A unique gift upgraded from a regular gift.

UniqueGiftBackdrop

Backdrop of a unique gift.

UniqueGiftBackdropColors

Colors of the backdrop of a unique gift.

UniqueGiftColors

Color scheme for a user’s name, replies, and link previews based on a unique gift.

UniqueGiftInfo

Service message: a unique gift was sent or received.

UniqueGiftModel

Model of a unique gift.

UniqueGiftSymbol

Symbol shown on the pattern of a unique gift.