rustigram_types::keyboard

Struct KeyboardButton

pub struct KeyboardButton {
    pub text: String,
    pub icon_custom_emoji_id: Option<String>,
    pub style: Option<ButtonStyle>,
    pub request_users: Option<KeyboardButtonRequestUsers>,
    pub request_chat: Option<KeyboardButtonRequestChat>,
    pub request_managed_bot: Option<KeyboardButtonRequestManagedBot>,
    pub request_contact: Option<bool>,
    pub request_location: Option<bool>,
    pub request_poll: Option<KeyboardButtonPollType>,
    pub web_app: Option<WebAppInfo>,
}

One button in a reply keyboard.

Fields
text: String

Label text on the button.

icon_custom_emoji_id: Option<String>

Custom emoji identifier shown before the button text.

style: Option<ButtonStyle>

Visual style of the button.

request_users: Option<KeyboardButtonRequestUsers>

Request to select and share one or more users.

request_chat: Option<KeyboardButtonRequestChat>

Request to select and share a chat.

request_managed_bot: Option<KeyboardButtonRequestManagedBot>

Request a managed bot from the user.

request_contact: Option<bool>

Requests the user’s phone number.

request_location: Option<bool>

Requests the user’s current location.

request_poll: Option<KeyboardButtonPollType>

Requests the user to create a poll.

web_app: Option<WebAppInfo>

Web App to launch when the button is pressed.

Implementations

impl KeyboardButton

pub fn text(label: impl Into<String>) -> Self

Creates a simple text button.

pub fn request_contact(label: impl Into<String>) -> Self

Creates a button that requests the user’s phone number.

pub fn request_location(label: impl Into<String>) -> Self

Creates a button that requests the user’s location.