rustigram_types::update

Enum UpdateKind

pub enum UpdateKind {

    Message(Message),
    EditedMessage(Message),
    ChannelPost(Message),
    EditedChannelPost(Message),
    BusinessMessage(Message),
    EditedBusinessMessage(Message),
    InlineQuery(InlineQuery),
    ChosenInlineResult(ChosenInlineResult),
    CallbackQuery(Box<CallbackQuery>),
    ShippingQuery(ShippingQuery),
    PreCheckoutQuery(PreCheckoutQuery),
    Poll(Poll),
    PollAnswer(PollAnswer),
    MyChatMember(ChatMemberUpdated),
    ChatMember(ChatMemberUpdated),
    ChatJoinRequest(ChatJoinRequest),
    MessageReaction(MessageReactionUpdated),
    MessageReactionCount(MessageReactionCountUpdated),
    ChatBoost(ChatBoostUpdated),
    RemovedChatBoost(ChatBoostRemoved),
    ManagedBot(ManagedBotUpdated),
    BusinessConnection(BusinessConnection),
    DeletedBusinessMessages(BusinessMessagesDeleted),
    PurchasedPaidMedia(PaidMediaPurchased),
    GuestMessage(Message),
    Subscription(BotSubscriptionUpdated),

}

The specific kind of an incoming update.

Variants

Message(Message)

New incoming message.

EditedMessage(Message)

New version of a message.

ChannelPost(Message)

New incoming channel post.

EditedChannelPost(Message)

New version of a channel post.

BusinessMessage(Message)

A message from a business account.

EditedBusinessMessage(Message)

Edited message from a business account.

InlineQuery(InlineQuery)

New incoming inline query.

ChosenInlineResult(ChosenInlineResult)

The result of an inline query that was chosen.

CallbackQuery(Box<CallbackQuery>)

New incoming callback query.

ShippingQuery(ShippingQuery)

New incoming shipping query (only for invoices with flexible price).

PreCheckoutQuery(PreCheckoutQuery)

New incoming pre-checkout query.

Poll(Poll)

New poll state.

PollAnswer(PollAnswer)

A user changed their answer in a non-anonymous poll.

MyChatMember(ChatMemberUpdated)

Bot’s chat member status was updated in a chat.

ChatMember(ChatMemberUpdated)

A chat member’s status was updated in a chat.

ChatJoinRequest(ChatJoinRequest)

A request to join the chat has been sent.

MessageReaction(MessageReactionUpdated)

A reaction to a message was changed by a user.

MessageReactionCount(MessageReactionCountUpdated)

Reactions to a message with anonymous reactions were changed.

ChatBoost(ChatBoostUpdated)

A chat boost was added or changed.

RemovedChatBoost(ChatBoostRemoved)

A boost was removed from a chat.

ManagedBot(ManagedBotUpdated)

A managed bot was connected or disconnected.

BusinessConnection(BusinessConnection)

A business connection was established or removed.

DeletedBusinessMessages(BusinessMessagesDeleted)

Messages were deleted from a connected business account.

PurchasedPaidMedia(PaidMediaPurchased)

Purchased paid media.

GuestMessage(Message)

New guest message — the bot can reply using answerGuestQuery.

Subscription(BotSubscriptionUpdated)

User payment subscription toward the bot has changed.